[blob,extras] move blob functions to the blob file
So that we avoid a level of indirection and slim down extras even
more.
