+++ accelerator/control.py
@@ -32,12 +32,12 @@ from accelerator import dispatch
 from accelerator import workspace
 from accelerator import database
 from accelerator import methods
+from accelerator.blob import json_save
 from accelerator.colourwrapper import colour
 from accelerator.compat import FileNotFoundError
 from accelerator.setupfile import update_setup
 from accelerator.job import WORKDIRS, Job
 from accelerator.error import BuildError
-from accelerator.extras import json_save
 from accelerator.types import DotDict
